Create Checkers.js
Create a checkerboard using a nested loop. Recall that a checkerboard is an 8x8 grid of alternating black and white squares, starting with a white in the top left corner. Draw red and green circles as "checkers" on the top two rows and the bottom two rows. You must use variables and control structures in the final version of your sketch.