Create recoding_schotter.js
Schotter by Georg Nees (1966) is a classic algorithmic artwork that depicts a gradient from order to chaos across a 12x22 grid of squares. In it, the squares' orientations become increasingly randomized towards the bottom of the page. Re-code this work, paying attention to detail. (https://digitalartarchive.siggraph.org/artwork/georg-nees-gravel-stones/